requirements.txt
coverage==4.5.1
dj-database-url==0.5.0
Django==2.1
django-filter==2.0.0
django-nose==1.4.5
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.2.1
drf-writable-nested==0.4.3
factory-boy==2.11.1
Faker==0.9.0
nose==1.3.7
Pillow==5.2.0
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
